requirements.txt
langchain>=0.1.0
langchain-google-genai>=0.0.6
langchain-community>=0.0.10
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
pypdf>=3.17.0
python-docx>=1.1.0
docx2txt>=0.8
streamlit>=1.29.0
python-dotenv>=1.0.0
filelock>=3.13.0
python-magic-bin>=0.4.14 ; platform_system == "Windows"
python-magic>=0.4.27 ; platform_system != "Windows"
pytest>=7.4.0
